Render pokemon when generation is clicked
